stocks/GT-requirements.txt
flask
flask_pymongo
requests
os



